ADDED: New draft chapter on artificial intelligence (AI) with Prolog.
This chapter was suggested by Anton Kochkov (@XVilka) in issue #5.
The Metagol reference was suggested by Stassa Patsantzis (@stassa)
in a HN discussion (https://news.ycombinator.com/item?id=17121028)
of The Power of Prolog. Many thanks!

This chapter will be updated with more use cases as they become
available. I am very interested in the chatbot example, RITS, and in
the formalisation of legislation. Please contact me if you want to
contribute to AI developments with Prolog. See also issue #5. I invite
you to add interesting suggestions and comments to that issue too.