Add a few videos.
The videos are work in progress and may be replaced by newer versions
at any time. Hence, if possible, please use the stable links to refer
to these videos, since they will always point to the latest version.

For example, to link to the Sudoku video, please use:

    https://www.metalevel.at/prolog/videos/sudoku

This will automatically redirect to the latest version. See also #3.