use Scryer Prolog for the DCG chapter
This has become possible due to the work of @mthom and @UWN,
culminating in the first Prolog system that efficiently represents
lists of characters in memory, and which now provides all features to
run every example in this chapter, including phrase_from_file/2 and
char_type/2 which have been brilliantly implemented and generously
contributed to Scryer Prolog by @notoria.

It is nice to finally demonstrate what Prolog was designed for:
Parsing text conveniently and efficiently.

Many thanks!